refactor: move the V1 writer into a test fixture file
v1.5.0 closed V1 to writes: v1Operations.add and remove return
ErrV1ReadOnly before looking at their arguments, so no released binary
can reach writeKeyword, deleteKeyword, or the per-character trie walk
they drove. All of it stayed in production anyway, reachable only from
tests.

It is moved rather than deleted because the fixtures need it. Seeding a
V1 collection by hand would mean a second copy of the trie-building
algorithm, and a copy that drifted would let the read and migration
paths pass against data no release ever wrote. Keeping the real writer
keeps those tests honest.

_test.go is what makes the guarantee structural: no production binary
carries a way to write V1, and the code cannot drift back onto a
reachable path without someone moving it out again.

trie.go is gone. Its one surviving function, outputWithContext, moved
next to debugV1 - its only remaining caller. The buildTrieHook field on
AhoCorasick and the v1Operations.ac back-pointer, whose comment called
it a temporary bridge until T15, went with it: the hook now belongs to
the fixture writer and is installed through setV1BuildTrieHook.

Everything moved is unexported, so api/v1.txt is byte-identical.
RollbackTimeout stays: the V1 flush still runs under it.

api/v1-audit.txt citations into v1_ops.go are repointed - the four the
gate caught, and the six it could not because they still resolved to a
line inside the shortened file.
