+++ changes/unreleased/Changed-20260809-140000.yaml
@@ -1,0 +1,5 @@
+kind: Changed
+body: 'Moved the V1 writer out of the production build and into `v1_fixture_test.go`. Since v1.5.0 closed V1 to writes, `v1Operations.add` and `remove` return `ErrV1ReadOnly` before looking at anything, so nothing reachable from a released binary could call `writeKeyword`, `deleteKeyword`, or the per-character trie walk they drove — `buildTrie`, `rebuildOutput`, `buildOutput`, `pruneTrie`, and the goto/fail helpers in `trie.go`. All of it survived only so tests could seed a collection in the real V1 layout, which is worth keeping: a hand-written fixture would be a second copy of the trie-building algorithm, and a divergent copy would let the read and migration paths pass against data no release ever wrote. Living in a `_test.go` file makes that structural — no production binary now carries a way to write V1, and the code cannot drift back onto a reachable path without being moved out again. `trie.go` is gone; its one surviving function, `outputWithContext`, moved next to `debugV1`, its only remaining caller. The `AhoCorasick.buildTrieHook` field and the `v1Operations.ac` back-pointer (commented "temporary bridge until T15") went with it: the hook now lives on the fixture writer and is installed through `setV1BuildTrieHook`. Everything moved is unexported, so `api/v1.txt` is byte-identical; `AhoCorasickArgs.RollbackTimeout` stays, since the V1 flush still runs under it.'
+time: 2026-08-09T14:00:00.000000+09:00
+custom:
+    Issue: "221"
